fix: game service, socket service 분리
