+++ server/src/main/java/com/wordrelay/server/controller/GameController.java
@@ -1,0 +1,18 @@
+package com.wordrelay.server.controller;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.messaging.handler.annotation.MessageMapping;
+import org.springframework.stereotype.Controller;
+
+@Controller
+@Slf4j
+@RequiredArgsConstructor
+public class GameController {
+
+
+
+
+
+
+}
