+++ server/src/main/java/com/wordrelay/server/controller/SocketController.java
@@ -1,6 +1,7 @@
 package com.wordrelay.server.controller;
 
 import com.wordrelay.server.service.GameService;
+import com.wordrelay.server.service.SocketService;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -11,21 +12,15 @@ import org.springframework.stereotype.Controller;
 @Controller
 @Slf4j
 @RequiredArgsConstructor
-public class GameWebSocketController {
-
-  private final GameService gameService;
+public class SocketController {
+  private final SocketService socketService;
 
   @MessageMapping("/game.connect")
   public void handleConnection(SimpMessageHeaderAccessor headerAccessor, Map<String, Object> payload) {
     String browserId = (String) payload.get("browserId");
-    String nickname = gameService.handleUserConnection(browserId);
+    String nickname = socketService.handleUserConnection(browserId);
     headerAccessor.getSessionAttributes().put("username", nickname);
-    log.info("User connected: " + nickname);
 
   }
-
-
-
-
 
 }
